kafkaGithubConsumer/requirements.txt
bytewax
orjson
pydantic
PyGithub
icecream
confluent-kafka
kafka-python-ng
aiohttp==3.9.0b0
pydantic-settings
pytest